change EventType ID generation (now vai typeid)

+++ EntityComponentSystem/include/Platform.h
@@ -83,8 +83,8 @@ namespace ECS
 	using iptr			= intptr_t;
 	using uptr			= uintptr_t;
 
-	using ObjectID		= u64;
-	using TypeID		= u64;
+	using ObjectID		= size_t;
+	using TypeID		= size_t;
 
 	static const ObjectID	INVALID_OBJECT_ID	= std::numeric_limits<ObjectID>::max();
 	static const TypeID		INVALID_TYPE_ID		= std::numeric_limits<TypeID>::max();
